requirements.txt
pymdown-extensions>=9.4
pytest>=7
PyYAML>=6
beautifulsoup4==4.10
mkdocs[i18n]>=1.4
requests==2.27
pygments>=2.13
python-dotenv ~= 0.21
